Move AcademicYear and SemesterNumber into packages
