+++ src/test/java/seedu/resireg/testutil/SemesterBuilder.java
@@ -7,10 +7,10 @@ import java.util.Map;
 
 import seedu.resireg.model.allocation.Allocation;
 import seedu.resireg.model.room.roomtype.RoomType;
-import seedu.resireg.model.semester.AcademicYear;
-import seedu.resireg.model.semester.RoomRate;
 import seedu.resireg.model.semester.Semester;
-import seedu.resireg.model.semester.SemesterNumber;
+import seedu.resireg.model.semester.academicyear.AcademicYear;
+import seedu.resireg.model.semester.roomrate.RoomRate;
+import seedu.resireg.model.semester.semesternumber.SemesterNumber;
 
 public class SemesterBuilder {
